fix(editor): finish editor.

+++ src/main/java/com/clougence/schema/metadata/domain/rdb/mysql/MySqlColumn.java
@@ -46,7 +46,7 @@ public class MySqlColumn implements ColumnDef {
     private String     defaultCollationName;
     private String     defaultCharacterSetName;
     private Long       charactersMaxLength;
-    private Integer    bytesMaxLength;
+    private Long       bytesMaxLength;
 
     public String getName() {
         return this.name;
